Refactored: Moved subcommands to seperate files & Switched from main.swift to `@main`
