+++ QRGen/Command/ExpressibleByArgument.swift
@@ -1,0 +1,12 @@
+//
+//  ExpressibleByArgument.swift
+//  QRGen
+//
+//  Created by Max-Joseph on 17.01.23.
+//
+
+import Foundation
+import ArgumentParser
+
+
+typealias ArgumentEnum = ExpressibleByArgument & CaseIterable
